Fixes
Fixed an alignment issue where, when a wire is finished placing, it get's snapped to the target pin, this would result in a slight misalignment of the second to last point on the wire.
The fix now reroutes that point when finishing placement of a wire.

Removed 90 degree routing mode for now, until I am able to implement it later. Want to look at curved wires first.